[tests] Used integer values for topology expiration_time
The Topology.expiration_time is a PositiveIntegerField in Django
that can only store positive integer values in the database. However,
when we retrieve this value from the database using the
.get() method (for eg, in the handle_update_topology Celery task),
Django will truncate any decimal values, causing tests to fail.

To avoid this issue, we should ensure that we only use positive
integer values for expiration time in our tests.