[fix] Fix visualizer quirks with big topologies #176
The previous patch in #176 did an asynchronous request to fetch
NetJSONGraph data. If the request promise didn't resovle before
initialization of the graph (i.e. it took a long time for a response),
then the graph was initialized with the default graphConfig.
This caused the logic added to disable layoutAnimation to become
obsolete.
This patch fixes this issue by updating graphConfig in
NetJSONGraph.onLoad.

Related to #176 #noqa